+++ controller/vendor_meta.go
@@ -5,7 +5,6 @@ import (
 
 	"github.com/QuantumNous/new-api/common"
 	"github.com/QuantumNous/new-api/model"
-	"github.com/QuantumNous/new-api/service"
 
 	"github.com/gin-gonic/gin"
 )
@@ -79,7 +78,6 @@ func CreateVendorMeta(c *gin.Context) {
 		common.ApiError(c, err)
 		return
 	}
-	service.InvalidateCCSwitchModelCache()
 	common.ApiSuccess(c, &v)
 }
 
@@ -107,7 +105,6 @@ func UpdateVendorMeta(c *gin.Context) {
 		common.ApiError(c, err)
 		return
 	}
-	service.InvalidateCCSwitchModelCache()
 	common.ApiSuccess(c, &v)
 }
 
@@ -123,6 +120,5 @@ func DeleteVendorMeta(c *gin.Context) {
 		common.ApiError(c, err)
 		return
 	}
-	service.InvalidateCCSwitchModelCache()
 	common.ApiSuccess(c, nil)
 }